change how feedback after editing/adding places is displaied to user

+++ app/Http/Controllers/PlaceRemoveController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use App\Models\Place;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PlaceRemoveController extends Controller
 {
@@ -14,7 +16,12 @@ class PlaceRemoveController extends Controller
 
     public function delete(Place $place)
     {
-        Place::destroy($place->id);
-        return back();
+        $user = User::find(Auth::id());
+        if ($user->hasRole('editor') || ($place->user_id == $user->id)) {
+            Place::destroy($place->id);
+            return back()->with('success', 'Place removed successfully.');
+        } else {
+            return back()->with('error', 'Unable to proceed with this request...');
+        }
     }
 }
